Move section padding calculator to `section.py`
